functions/add_watermark/requirements.txt
functions-framework==3.*
google-cloud-speech
google-cloud-aiplatform
google-cloud-storage
moviepy
pyffmpeg
ffmpeg